auth tests works fixing users after

+++ src/controllers/usersController.js
@@ -1,37 +1,49 @@
 const User = require("../models/User");
 
-exports.update = async (req, res) => {     
-  try {
-    const updatedUser = await User.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: req.body,
-      },
-      { new: true }
-    );
-    const { password, ...info } = updatedUser._doc;
-    res.status(200).json(info);
-  } catch (err) {
-    res.status(500).json(err);
+exports.update = async (req, res) => {   
+  if(req.params.id && (req.body.username || req.body.password || req.body.email)){  
+    try {
+      const updatedUser = await User.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: req.body,
+        },
+        { new: true }
+      );
+      const { password, ...info } = updatedUser._doc;
+      res.status(200).json(info);
+    } catch (err) {
+      res.status(500).json(err);
+    }
+  }else{
+    res.status(400).json();
   }
 }
 
 exports.delete = async (req, res) => {
-  try {
-    await User.findByIdAndDelete(req.params.id);
-    res.status(200).json("User has been deleted...");
-  } catch (err) {
-    res.status(500).json(err);
+  if(req.params.id){
+    try {
+      await User.findByIdAndDelete(req.params.id);
+      res.status(200).json("User has been deleted...");
+    } catch (err) {
+      res.status(500).json(err);
+    }
+  }else{
+    res.status(400).json();
   }
 }
 
 exports.get = async (req, res) => {
-  try {
-    const user = await User.findById(req.params.id);
-    const { password, ...info } = user._doc;
-    res.status(200).json(info);
-  } catch (err) {
-    res.status(500).json(err);
+  if(req.params.id){
+    try {
+      const user = await User.findById(req.params.id);
+      const { password, ...info } = user._doc;
+      res.status(200).json(info);
+    } catch (err) {
+      res.status(500).json(err);
+    }
+  }else{
+    res.status(400).json();
   }
 }
